controller-frontend

+++ src/main/java/app/controllers/InquiryController.java
@@ -1,5 +1,4 @@
 package app.controllers;
-
 import app.dto.requestDTO.InquiryRequestDTO;
 import app.dto.requestDTO.RoofRequestDTO;
 import app.dto.requestDTO.carports.CarportRequestDTO;
@@ -8,6 +7,7 @@ import app.dto.responseDTO.UserResponseDTO;
 import app.exceptions.CalculatorException;
 import app.exceptions.DatabaseException;
 import app.services.ServiceFactory;
+import app.services.utils.UserValidator;
 import io.javalin.Javalin;
 import io.javalin.http.Context;
 import java.util.List;
@@ -15,13 +15,10 @@ import java.util.List;
 public class InquiryController {
 
     public void addRoutes(Javalin app, ServiceFactory serviceFactory) {
-        app.post("/createInquiry", ctx -> createInquiry(ctx, serviceFactory));
-        app.post("/getInquiry/{inquiry_id}", ctx -> getInquiry(ctx, serviceFactory));
-        app.post("/getAllInquiries", ctx -> getAllInquiries(ctx, serviceFactory));
-        app.post("/deleteInquiry", ctx -> deleteInquiry(ctx, serviceFactory));
-        app.get("/getInquiryCustomer", ctx -> ctx.render("customer-inquiry-details")); /* Test route af AJ*/
-        app.get("/getInquiryAdmin", ctx -> ctx.render("admin-inquiry-details")); /* Test route af AJ*/
-
+        app.post("/create/inquiry", ctx -> createInquiry(ctx, serviceFactory));
+        app.get("/customer/get/inquiry/{inquiry_id}", ctx -> customerGetInquiryDetails(ctx, serviceFactory));
+        app.get("/admin/get/inquiry/{inquiry_id}", ctx -> adminGetInquiryDetails(ctx, serviceFactory));
+        app.post("/admin/delete/inquiry", ctx -> adminDeleteInquiry(ctx, serviceFactory));
     }
 
     public void createInquiry(Context ctx, ServiceFactory serviceFactory) throws CalculatorException, DatabaseException {
@@ -37,11 +34,28 @@ public class InquiryController {
         RoofRequestDTO roofRequestDTO = new RoofRequestDTO(roofSlope, roofMaterial, roofType);
 
         //Shed
-        String shedWidth = ctx.formParam("shed_width");
+        String shedWidth;
+        String shedStatus = ctx.formParam("shed_status");
         String shedLength = ctx.formParam("shed_length");
         String shedSiding = ctx.formParam("shed_siding");
-        boolean floor = Boolean.parseBoolean(ctx.formParam("floor"));
+        String floorStatus = ctx.formParam("shed_floor");
+        boolean floor;
 
+        if (shedStatus.matches("HALF")){
+            shedWidth = String.valueOf(carportWidth/2);
+        } else if (shedStatus.matches("FULL")){
+            shedWidth = String.valueOf(carportWidth);
+        } else {
+            shedLength = null;
+            shedWidth = null;
+            shedSiding = null;
+            floor = false;
+        }
+        if (floorStatus.matches("TRUE")){
+            floor=true;
+        }else {
+            floor = false;
+        }
 
         CarportRequestDTO carportRequestDTO = serviceFactory.getCarportService().checkShed(shedWidth, shedLength, shedSiding, floor, carportWidth, carportHeight, carportLength, roofRequestDTO);
 
@@ -55,14 +69,14 @@ public class InquiryController {
             ctx.sessionAttribute("pending_roof_slope", ctx.formParam("roof_slope"));
             ctx.sessionAttribute("pending_roof_material", ctx.formParam("roof_material"));
             ctx.sessionAttribute("pending_roof_type", ctx.formParam("roof_type"));
-            ctx.sessionAttribute("pending_shed_width", ctx.formParam("shed_width"));
+            ctx.sessionAttribute("pending_shed_width", ctx.formParam("shed_status"));
             ctx.sessionAttribute("pending_shed_length", ctx.formParam("shed_length"));
             ctx.sessionAttribute("pending_shed_siding", ctx.formParam("shed_siding"));
-            ctx.sessionAttribute("pending_floor", ctx.formParam("floor"));
+            ctx.sessionAttribute("pending_floor", ctx.formParam("shed_floor"));
             ctx.sessionAttribute("pending_remark", ctx.formParam("inquiry_remark"));
             ctx.sessionAttribute("pending_inquiry", true);
 
-            ctx.redirect("/login");
+            ctx.redirect("/customer/login");
             return;
         }
 
@@ -72,31 +86,41 @@ public class InquiryController {
         InquiryRequestDTO inquiryRequestDTO = new InquiryRequestDTO(customerId, inquiryRemark, carportId);
         serviceFactory.getInquiryService().createInquiry(inquiryRequestDTO);
 
-        ctx.render("/profilePage");
+        ctx.redirect("/customer/my/page");
     }
 
-    public void getInquiry(Context ctx, ServiceFactory serviceFactory) throws DatabaseException, CalculatorException {
-        int inquiryId = Integer.parseInt(ctx.pathParam("inquiry_id"));
-        UserResponseDTO userResponseDTO = ctx.sessionAttribute("currentUser");
-        InquiryResponseDTO inquiryResponseDTO = serviceFactory.getInquiryService().getInquiry(inquiryId);
-
-        if (inquiryResponseDTO.getCustomerResponseDTO().getId() != userResponseDTO.getId()){
-            ctx.status(403);
+    public void customerGetInquiryDetails(Context ctx, ServiceFactory serviceFactory) throws DatabaseException, CalculatorException {
+        if (!UserValidator.isCustomer(ctx)){
+            ctx.redirect("/");
             return;
         }
+        int inquiryId = Integer.parseInt(ctx.pathParam("inquiry_id"));
+        InquiryResponseDTO inquiryResponseDTO = serviceFactory.getInquiryService().getInquiry(inquiryId);
+
+        ctx.sessionAttribute("inquiry_responseDTO",inquiryResponseDTO);
+        ctx.attribute("selected_inquiry", inquiryResponseDTO);
+        ctx.render("customer-inquiry-details.html");
+    }
+
+    public void adminGetInquiryDetails(Context ctx, ServiceFactory serviceFactory) throws DatabaseException, CalculatorException {
+        if (!UserValidator.isAdmin(ctx)){
+            ctx.redirect("/");
+            return;
+        }
+
+        int inquiryId = Integer.parseInt(ctx.pathParam("inquiry_id"));
+        InquiryResponseDTO inquiryResponseDTO = serviceFactory.getInquiryService().getInquiry(inquiryId);
 
         ctx.sessionAttribute("inquiry_responseDTO",inquiryResponseDTO);
         ctx.attribute("selected_inquiry", inquiryResponseDTO);
         ctx.render("admin-inquiry-details.html");
     }
 
-    public void getAllInquiries(Context ctx, ServiceFactory serviceFactory) throws DatabaseException {
-        List<InquiryResponseDTO> inquiryResponseDTOList = serviceFactory.getInquiryService().getAllInquiries();
-        ctx.attribute("all_inquiries", inquiryResponseDTOList);
-        ctx.render("admin-all-inquiries.html");
-    }
-
-    public void deleteInquiry(Context ctx, ServiceFactory serviceFactory) throws DatabaseException, CalculatorException {
+    public void adminDeleteInquiry(Context ctx, ServiceFactory serviceFactory) throws DatabaseException, CalculatorException {
+        if (!UserValidator.isAdmin(ctx)){
+            ctx.redirect("/");
+            return;
+        }
         int inquiryId = Integer.parseInt(ctx.pathParam("inquiry_id"));
         InquiryResponseDTO inquiryResponseDTO = serviceFactory.getInquiryService().getInquiry(inquiryId);
         serviceFactory.getInquiryService().deleteInquiry(inquiryId);
